refactor(reply): relocate quality gate implementation
